Adding connections to settings api's
Adding request to pull the github repos as well as name and age. This will need to be changed so that the data stays bound to the first load and then is updated periodically. There is no need to update this each time the component is run although we can.