Adding value & other optimizations
Ability to provide a default value is great, now that can be done. Additionally this no longer had a giant array built in it will now use a little compute power to do that client side since there is no need to include that in the app.js compiled form.